macos_defaults: don't trust an empty selection (#60)
Fixes https://github.com/phillco/talon-axkit/issues/59; see that issue
for context. tldr: VS Code has started (when?) to report some partial
accessibility information, which unfortunately only includes the
selected text sometimes. Sometimes it says it is empty when it isn't.

In the future I could see us having a setting to adjust this behavior
that could be turned on for fully native text fields that don't do what
VS Code is doing, but in the interim, I think this is reasonable?

Also remove a fallback for <0.3 that's no longer needed.